Refactoring: separate clients and move tests outside of the code
