+++ rosreestr_api/clients/rosreestr.py
@@ -1,135 +1,13 @@
-import time
 import logging
-from typing import Union
 from dataclasses import dataclass
-from urllib.parse import urlencode, quote_plus
+from urllib.parse import quote_plus
 
 import requests
 
+from rosreestr_api.clients.http import HTTPClient
+
 
 logger = logging.getLogger(__name__)
-
-
-def _strip_cadastral_id(cadastral_id):
-    stripped_cadastral_id = []
-    cadastral_id = cadastral_id.split(':')
-    for part in cadastral_id:
-        if part:
-            stripped_cadastral_id.append(part[:-1].lstrip('0') + part[-1])
-    return ':'.join(stripped_cadastral_id)
-
-
-def _get_body_for_logging(body: Union[bytes, str]) -> str:
-    try:
-        if isinstance(body, bytes):
-            return (b' BODY: ' + body).decode('utf-8')
-        elif isinstance(body, str):
-            return ' BODY: ' + body
-        else:
-            return ''
-    except UnicodeDecodeError:
-        return ''
-
-
-def _get_duration_for_logging(duration: str) -> str:
-    if duration is not None:
-        return ' {0:.6f}s'.format(duration)
-    else:
-        return ''
-
-
-class HTTPClient:
-
-    GET_HTTP_METHOD = 'GET'
-    POST_HTTP_METHOD = 'POST'
-    PATCH_HTTP_METHOD = 'PATCH'
-    PUT_HTTP_METHOD = 'PUT'
-
-    BODY_LESS_METHODS = [GET_HTTP_METHOD]
-    LOG_REQUEST_TEMPLATE = '%(method)s %(url)s%(request_body)s%(duration)s'
-    LOG_RESPONSE_TEMPLATE = (LOG_REQUEST_TEMPLATE +
-                             ' - HTTP %(status_code)s%(response_body)s%(duration)s')
-
-    def __init__(self, timeout=3, keep_alive=False, default_headers=None):
-        self.timeout = timeout
-        self.keep_alive = keep_alive
-        self.default_headers = default_headers or {}
-        self._session = None
-
-    def _log_request(self, method, url, body, duration=None, log_method=logger.info):
-        message_params = {
-            'method': method, 'url': url, 'request_body': _get_body_for_logging(body),
-            'duration': _get_duration_for_logging(duration)}
-        log_method(self.LOG_REQUEST_TEMPLATE, message_params)
-
-    def _log_response(self, response, duration, log_method=logger.info):
-        message_params = {
-            'method': response.request.method,
-            'url': response.request.url,
-            'request_body': _get_body_for_logging(response.request.body),
-            'status_code': response.status_code,
-            'response_body': _get_body_for_logging(response.content),
-            'duration': _get_duration_for_logging(duration)}
-        log_method(self.LOG_RESPONSE_TEMPLATE, message_params)
-
-    def _make_request(self, method, url, **kwargs) -> requests.Response:
-        kwargs.setdefault('timeout', self.timeout)
-        session = self.session
-        timeout = kwargs.pop('timeout', self.timeout)
-
-        headers = self.default_headers.copy()
-        headers.update(kwargs.pop('headers', {}))
-
-        request = requests.Request(method, url, headers=headers, **kwargs)
-        prepared_request = request.prepare()
-        self._log_request(method, url, prepared_request.body)
-        start_time = time.time()
-        try:
-            response = session.send(prepared_request, timeout=timeout)
-            duration = time.time() - start_time
-            if response.status_code >= 400:
-                log_method = logging.error
-            else:
-                log_method = logging.debug
-
-            self._log_response(response, duration=duration, log_method=log_method)
-            return response
-        except requests.exceptions.RequestException as e:
-            duration = time.time() - start_time
-            if e.response:
-                self._log_response(e.response, duration=duration, log_method=logging.error)
-            else:
-                self._log_request(method, url, prepared_request.body, log_method=logging.exception)
-            raise
-        finally:
-            if not self.keep_alive:
-                session.close()
-
-    @property
-    def session(self) -> requests.Session:
-        if self.keep_alive:
-            if not self._session:
-                self._session = requests.Session()
-            return self._session
-        else:
-            return requests.Session()
-
-    def get(self, url, params=None, **kwargs) -> requests.Response:
-        if params:
-            url_with_query_params = url + '?' + urlencode(params)
-        else:
-            url_with_query_params = url
-
-        return self._make_request(self.GET_HTTP_METHOD, url_with_query_params, **kwargs)
-
-    def post(self, url, **kwargs) -> requests.Response:
-        return self._make_request(self.POST_HTTP_METHOD, url, **kwargs)
-
-    def patch(self, url, **kwargs) -> requests.Response:
-        return self._make_request(self.PATCH_HTTP_METHOD, url, **kwargs)
-
-    def put(self, url, **kwargs) -> requests.Response:
-        return self._make_request(self.PUT_HTTP_METHOD, url, **kwargs)
 
 
 @dataclass
@@ -156,7 +34,7 @@ class AddressWrapper:
 
 class RosreestrAPIClient:
 
-    BASE_URL = 'http://rosreestr.ru/api/online'
+    BASE_URL = 'https://rosreestr.gov.ru/api/online'
     MACRO_REGIONS_URL = f'{BASE_URL}/macro_regions/'
     REGIONS_URL = f'{BASE_URL}/regions/' + '{}/'
     REGION_TYPES_URL = f'{BASE_URL}/region_types/' + '{}/'
@@ -317,3 +195,12 @@ class PKKRosreestrAPIClient:
         url = self.SEARCH_BUILDING_BY_COORDINATES_URL.format(
             lat=lat, long=long, limit=limit, tolerance=tolerance)
         return self._http_client.get(url).json()
+
+
+def _strip_cadastral_id(cadastral_id):
+    stripped_cadastral_id = []
+    cadastral_id = cadastral_id.split(':')
+    for part in cadastral_id:
+        if part:
+            stripped_cadastral_id.append(part[:-1].lstrip('0') + part[-1])
+    return ':'.join(stripped_cadastral_id)
