+++ tests/test_api_clients.py
@@ -3,7 +3,7 @@ from unittest.mock import patch, MagicMock
 import pytest
 import httpretty
 
-from rosreestr_api.tests import rosreestr_client_fixtures, pkk_client_fixtures
+from tests import pkk_client_fixtures, rosreestr_client_fixtures
 from rosreestr_api.clients import (
     AddressWrapper, RosreestrAPIClient, PKKRosreestrAPIClient)
 
@@ -60,10 +60,12 @@ class TestRosreestrAPIClient:
             method=httpretty.GET, uri=self.MACRO_REGIONS_URL, body=rosreestr_client_fixtures.MACRO_REGIONS_RESPONSE,
             content_type=self.CONTENT_TYPE_JSON)
         httpretty.register_uri(
-            method=httpretty.GET, uri=self.REGIONS_URL.format(rosreestr_client_fixtures.MACRO_REGION_ID_1),
+            method=httpretty.GET, uri=self.REGIONS_URL.format(
+                rosreestr_client_fixtures.MACRO_REGION_ID_1),
             body=rosreestr_client_fixtures.MACRO_REGION_TO_REGION_1_RESPONSE, content_type=self.CONTENT_TYPE_JSON)
         httpretty.register_uri(
-            method=httpretty.GET, uri=self.REGIONS_URL.format(rosreestr_client_fixtures.MACRO_REGION_ID_2),
+            method=httpretty.GET, uri=self.REGIONS_URL.format(
+                rosreestr_client_fixtures.MACRO_REGION_ID_2),
             body=rosreestr_client_fixtures.MACRO_REGION_TO_REGION_2_RESPONSE, content_type=self.CONTENT_TYPE_JSON)
 
         api_client = RosreestrAPIClient()
@@ -73,12 +75,14 @@ class TestRosreestrAPIClient:
     @httpretty.activate
     def test_get_region_types(self):
         httpretty.register_uri(
-            method=httpretty.GET, uri=self.REGION_TYPES_URL.format(rosreestr_client_fixtures.REGION_ID),
+            method=httpretty.GET, uri=self.REGION_TYPES_URL.format(
+                rosreestr_client_fixtures.REGION_ID),
             body=rosreestr_client_fixtures.REGION_TYPES_RESPONSE, content_type=self.CONTENT_TYPE_JSON)
 
         api_client = RosreestrAPIClient()
 
-        assert rosreestr_client_fixtures.REGION_TYPES == api_client.get_region_types(rosreestr_client_fixtures.REGION_ID)
+        assert rosreestr_client_fixtures.REGION_TYPES == api_client.get_region_types(
+            rosreestr_client_fixtures.REGION_ID)
 
     @httpretty.activate
     def test_get_objects_by_right(self):
